Correctly set up and cleanup home view
The home view uses the browse app but wasn't correctly
setting up and tearing down the home view.  The links
weren't working either, when you'd click on a project.

This commit moves the browse controller to the #container
and home view to #home.  Keeps them seperate, and makes
the navigation work.

Fixes #2.