Removed project refs.
Removed additional references to projects from the home page and
about page if the UI configuration is set to not use projects.

Also combined the customizations that used to be separated into
another repo, to simplify maintenance. This was mostly due to
the addition of a home page in the main midas codebase.